Outsourced the parser to a different folder
